Upgrade node to v18 and remove babel
First order of business is to get right of the babel transpiling, which
serves literally no purpose. Removing babel removes the build step
entirely, and the only change that we have to make is to add file
extensions to our imports, and fix the imports on a handful of
libraries. I tested this against the frontend and everything looked
fine, but it's possible that a PUT route touches something that I didn't
import properly. Fortunately that's a trivial fix if it does come up.

Along the same lines, upgrade to the latest LTS version of node (18), so
that we can take advantage of various features that it provides,
including the test runner. I also switched to `npm`, which I guess is a
personal preference, but I am skeptical that there is any good reason to
use a non-default package manager in 2022.