Fix leader drop logic
Because the club leader drop logic only checked whether the user had
fewer clubs than before, it was possible for users to specify fewer
clubs, but different ones, and then bypass the approval stage for
becoming a leader in those clubs. I fixed this by verifying that the new
set of clubs was a subset of the previous one.

Resolves: #3