Add TODO messages for decomposing async + promise confusion
There are a few functions that return promises with async functions
inside them that should definitely be decomposed into a single,
top-level async function. I think it would be a little reckless for me
to do without a test of the functionality in place, so I'm just marking
it down for now so we can get the linter passing.