Optimize public trips route
First strike in the war against extremely slow API calls. I'm going to
be eliding mongoose entirely so that the queries are simpler to
translate to SQL later on. I will also be filtering the data so that
only the data presented to the view actually gets send to the frontend.
Safer and faster.