Add /leaders route for logged-in users
This is paired with a UI change to populate a leaders dropdown. The way
this worked previously was eye-popping: it just asked for every user,
the API would send the entirety of every user's data, and then the UI
would just filter that slightly.

Now we send a list of leaders and only include their email, name, and
ID. That's it! You have to be logged in to GET the route.