Add new /bookings bookings route
The old getVehicles function actually got us a list of vehicle bookings,
which is very different from when you need just a list of vehicles. So
make this route give the old "big" data response (for now) and then we
can move the vehicle dropdowns to a new, slimmer response that doesn't
populate with a ton of extraneous information.