Set the scheduled emails to not overlap
We're getting throttled by our email provider for having too many
simoultaneous connections. These mail functions are definitely blocking
the event loop so I'll rewrite them ASAP, but in the meantime they do
not need to be run as frequently as they are, so I'm decreasing the
frequency and scheduling them to not overlap.