Rewrite email senders to execute in order and without blocking
While there are a few event loop blocker in the application, this was
the big disaster. Four jobs ran *every minute* that each looped through
the entire corpus of trips, synchronously.

Fixing these required writing more specific queries that returned fewer
trips to filter through. There are also more async functions in each
loop, even though the loop processes each email in order. It's totally
fine to query mongo multiple times to fulfill these jobs; that's closer
to the patter we're going to be adopting with SQL anyway.

While I was going through it, I broke down the jobs into a more
composable style, with a mail sender that handles the mail formats and
transmission and controllers that fetch the user info. This was
especially necessary to ensure that the in-order sending logic was
replicated across each type of email.