Build more resilient club list enhancer
The list of clubs that's actually stored in MongoDB can be one of three
things: a list of strings, a list of ObjectIDs, and a list of clubs.
I built an enhance that is resilient to all three of these possibilities
(even though it's obviously a larger problem that needs to be solved).