Simplify DELETE /vehiclerequests route
This is an important one. The old route would call this mammoth
recompute function, that recomputed the conflicts between all the
vehicle requests in the database. As it turns out, the only time that
this total recomputation happens is when a vehicle is deleted, and you
can instead just remove that vehicle request from the conflict lists of
all the other vehicle requests.

This is one of the last major event loop blockers!