Update vehicle request saving route
Basically trim all the extraneous promises. I also remove the booking
data on the vehicle (which we did not use) and the conflict data (which
we did use but in a sufficiently minimal way that it was not worth the
complexity).

Conflict checking should be added back in a far simpler way in the
future.