Add "other costs" back to Pcard request
This was a signficant regression from the old Trailhead but I put off
re-implementing it for a while because I knew it would be annoying. And
it was in fact very annoying. I think I basically got it... we'll see.

I did have to remove the confirmation on saving Pcards with no other
costs, which is unfortunate (since the the target is different depending
on whether or not you added an other cost) but it won't be too hard to
add back. On the plus side, the leader view Pcard table works a little
more intuitively now.

Resolves: #30