Add new test framework
This script starts up the whole node server and then runs fetch commands
against it. The script also takes care of cleaning up the node sever
when its done.

While testing over the HTTP boundary has its pros and cons,
I think that it's the right level of granularity for this particular
application at this point because what I really want to be able to test
is a lot of stuff that's baked into express, like "can the right users
access this?" and "does this properly turn into HTML?"