Serve JS dependencies directly with long cache times
We have two client-side JS dependencies right now: HTMX, which we use
everywhere, and fullcalendar, which only the OPO view uses. We vend them
directly (instead of from a CDN) because it's more secure, but in theory
you don't want your node server wasting its time on static files, so
there's an NGINX layer in front of it.

This commit sets up a caching systems where these dependencies are
vended at URLs that include their version numbers, with a year-long
cache time. Serving the files once is no big deal, and if we need to
upgrade them, it will automatically change the URL (by parsing the
package.json file) and then all the users' browsers will request the
file again.

I think it's pretty clever!

Resolves: #73