Move the email-building code into its own function
Previously, code to send an email combined database queries, email
templating, and the actual mailer. This commit makes the database an
argument to the function that templates the email, and separates out the
mailer entirely. This will make it possible to test the email-building
functionality while retaining the convience of allowing the
email-builders to still query the database.

Right now the "database" that's passed is just a reference to the sqlite
module, but what's coming next is that the database connection will be
passed as a parameter to the server, which will let us test various
routes of the database in parallel, in isolation, and contained within
the node test runner.