Inject the Trailhead database as a dependency
This is a tricky one but the basic change here is that the database
connection is initialized and passed to the server at startup, rather
than imported directly and initialized as part of the module singleton.
This allows us to write really fast server tests at the http boundary,
because we can initialized multiple servers in parallel that have
different in-memory SQLite databases.

The drawback is that you use the intellisense on the module and some of
the linting benefits. Since most of the interactions with the database
come through get/all SQL methods, I think that's a totally fine
tradeoff. Some of the other extraneous helper methods could probably be
refactored out of the connector anyway.