Switch to using NetID for student emails
The new Microsoft email API will refuse to send any emails if even one
of the recipients is malformatted (great API Microsoft!).

This has prompted me to do a harsher version of a migration I was going
to do anyway: send emails to users based on their NetIDs, not the email
that they input. I would have like to preserve the functionality to let
them choose an email, but that requires writing validation code that I
think would be tricky to do correctly, and if I get it wrong, everyone
on the trip doesn't get emails.

They can learn how to set up forwarding if they want.